requirements.txt
ipython
numpy
torchvision